requirements.txt
frozendict>=1.2
mock
click>=6.6
Flask>=0.11.1
itsdangerous>=0.24
Jinja2>=2.8
jsonschema>=2.5.1
MarkupSafe>=0.23
PyYAML>=3.11
simplejson>=3.16.0
SQLAlchemy>=1.0.13
Werkzeug>=0.11.10
Sphinx>=1.5.1
